requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
attrs==24.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blis==1.2.0
catalogue==2.0.10
cbor==1.0.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.3.1
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.16.1
FlagEmbedding==1.3.4
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
ijson==3.3.0
inscriptis==2.5.3
ipykernel==6.29.5
ipython==8.32.0
ir_datasets==0.5.9
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.35
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-milvus==0.1.7
langchain-openai==0.3.6
langchain-text-splitters==0.3.4
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.2.4
language_data==1.3.0
lxml==5.3.1
lz4==4.4.3
mariadb==1.1.12
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.10
minio==7.2.15
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
newmm-tokenizer==0.2.2
nltk==3.9.1
numpy==1.26.4
onnxruntime==1.20.1
openai==1.61.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
peft==0.14.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.19.1
pymilvus==2.5.1
pymilvus.model==0.3.0
PyMuPDF==1.25.3
pyodbc==5.2.0
pyparsing==3.2.1
pypdfium2==4.30.1
pythainlp==5.0.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.1
ragas==0.2.13
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
semantic-router==0.0.72
semantic-text-splitter==0.23.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trec-car-tools==2.6
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
unlzw3==0.2.3
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zlib-state==0.1.9
